test(api): share edge cache fixtures across seasonal reads
